Go to page based on dynamic routing
