+++ pages/app/accounts/index.tsx
@@ -24,6 +24,7 @@ import CreditCardModal from "../../../components/modals/AccountModals/CreditCard
 import FixedInvestmentsModal from "../../../components/modals/AccountModals/FixedInvestmentsModal";
 import OtherAssetsModal from "../../../components/modals/AccountModals/OtherAssetsModal";
 import { BudgetEngineUtils } from "../../../src/engine/BudgetEngineUtils";
+import Link from "next/link";
 
 export default function AccountsPage() {
   const { useRequiredAuth } = useAuth();
@@ -131,38 +132,46 @@ export default function AccountsPage() {
                   const currAccount =
                     userData.financialInfo.accounts.bankAccounts[accountKey];
                   return (
-                    <Card
-                      bgColor={"white"}
+                    <Link
+                      href={`/app/accounts/${currAccount.name
+                        .toLowerCase()
+                        .split(" ")
+                        .join("-")}`}
                       key={accountKey}
-                      justify="space-between"
                     >
-                      <CardBody>
-                        <Badge>{currAccount.type}</Badge>
-                        <Heading size="sm"> {currAccount.name} </Heading>
-                        <Stat>
-                          <StatLabel>Account Value</StatLabel>
-                          <StatNumber>${currAccount.value}</StatNumber>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Account Interest Rate</StatLabel>
-                          <StatNumber>{currAccount.interestRate}%</StatNumber>
-                        </Stat>
-                        <Button
-                          onClick={() => {
-                            deleteAccount(
-                              userData.uid,
-                              userData.financialInfo.accounts,
-                              "BankAccount",
-                              accountKey
-                            );
-                          }}
-                          size="xs"
-                          variant="link"
-                        >
-                          Delete
-                        </Button>
-                      </CardBody>
-                    </Card>
+                      <a>
+                        <Card bgColor={"white"} justify="space-between">
+                          <CardBody>
+                            <Badge>{currAccount.type}</Badge>
+                            <Heading size="sm"> {currAccount.name} </Heading>
+                            <Stat>
+                              <StatLabel>Account Value</StatLabel>
+                              <StatNumber>${currAccount.value}</StatNumber>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Account Interest Rate</StatLabel>
+                              <StatNumber>
+                                {currAccount.interestRate}%
+                              </StatNumber>
+                            </Stat>
+                            <Button
+                              onClick={() => {
+                                deleteAccount(
+                                  userData.uid,
+                                  userData.financialInfo.accounts,
+                                  "BankAccount",
+                                  accountKey
+                                );
+                              }}
+                              size="xs"
+                              variant="link"
+                            >
+                              Delete
+                            </Button>
+                          </CardBody>
+                        </Card>
+                      </a>
+                    </Link>
                   );
                 }
               )}
@@ -174,61 +183,69 @@ export default function AccountsPage() {
                       accountKey
                     ];
                   return (
-                    <Card
-                      bgColor={"white"}
+                    <Link
+                      href={`/app/accounts/${currAccount.name
+                        .toLowerCase()
+                        .split(" ")
+                        .join("-")}`}
                       key={accountKey}
-                      justify="space-between"
                     >
-                      <CardBody>
-                        <Badge>FIXED INVESTMENT</Badge>
-                        <Heading size="sm"> {currAccount.name} </Heading>
-                        <Stat>
-                          <StatLabel>Account Value</StatLabel>
-                          <Stat>
-                            <StatLabel>Start date</StatLabel>
-                            <StatLabel>
-                              {(currAccount.startDate as Timestamp)
-                                .toDate()
-                                .toISOString()
-                                .substring(0, 10)}
-                            </StatLabel>
-                          </Stat>
-                          <Stat>
-                            <StatLabel>Maturity date</StatLabel>
-                            <StatLabel>
-                              {(currAccount.maturityDate as Timestamp)
-                                .toDate()
-                                .toISOString()
-                                .substring(0, 10)}
-                            </StatLabel>
-                          </Stat>
-                          <Stat>
-                            <StatLabel>Starting value</StatLabel>
-                            <StatNumber>
-                              ${currAccount.startingValue}
-                            </StatNumber>
-                          </Stat>
-                          <Stat>
-                            <StatLabel>Interest rate</StatLabel>
-                            <StatNumber>{currAccount.interestRate}%</StatNumber>
-                          </Stat>
-                        </Stat>
-                        <Button
-                          onClick={() => {
-                            deleteAccount(
-                              userData.uid,
-                              userData.financialInfo.accounts,
-                              "FixedInvestments",
-                              accountKey
-                            );
-                          }}
-                          size="xs"
-                          variant="link"
-                        >
-                          Delete
-                        </Button>
-                      </CardBody>
-                    </Card>
+                      <a>
+                        <Card bgColor={"white"} justify="space-between">
+                          <CardBody>
+                            <Badge>FIXED INVESTMENT</Badge>
+                            <Heading size="sm"> {currAccount.name} </Heading>
+                            <Stat>
+                              <StatLabel>Account Value</StatLabel>
+                              <Stat>
+                                <StatLabel>Start date</StatLabel>
+                                <StatLabel>
+                                  {(currAccount.startDate as Timestamp)
+                                    .toDate()
+                                    .toISOString()
+                                    .substring(0, 10)}
+                                </StatLabel>
+                              </Stat>
+                              <Stat>
+                                <StatLabel>Maturity date</StatLabel>
+                                <StatLabel>
+                                  {(currAccount.maturityDate as Timestamp)
+                                    .toDate()
+                                    .toISOString()
+                                    .substring(0, 10)}
+                                </StatLabel>
+                              </Stat>
+                              <Stat>
+                                <StatLabel>Starting value</StatLabel>
+                                <StatNumber>
+                                  ${currAccount.startingValue}
+                                </StatNumber>
+                              </Stat>
+                              <Stat>
+                                <StatLabel>Interest rate</StatLabel>
+                                <StatNumber>
+                                  {currAccount.interestRate}%
+                                </StatNumber>
+                              </Stat>
+                            </Stat>
+                            <Button
+                              onClick={() => {
+                                deleteAccount(
+                                  userData.uid,
+                                  userData.financialInfo.accounts,
+                                  "FixedInvestments",
+                                  accountKey
+                                );
+                              }}
+                              size="xs"
+                              variant="link"
+                            >
+                              Delete
+                            </Button>
+                          </CardBody>
+                        </Card>
+                      </a>
+                    </Link>
                   );
                 }
               )}
@@ -240,34 +257,40 @@ export default function AccountsPage() {
                   const currAccount =
                     userData.financialInfo.accounts.otherAssets[accountKey];
                   return (
-                    <Card
-                      bgColor={"white"}
+                    <Link
+                      href={`/app/accounts/${currAccount.name
+                        .toLowerCase()
+                        .split(" ")
+                        .join("-")}`}
                       key={accountKey}
-                      justify="space-between"
                     >
-                      <CardBody>
-                        <Badge>OTHER ASSET</Badge>
-                        <Heading size="sm"> {currAccount.name} </Heading>
-                        <Stat>
-                          <StatLabel>Asset Value</StatLabel>
-                          <StatNumber>${currAccount.value}</StatNumber>
-                        </Stat>
-                        <Button
-                          onClick={() => {
-                            deleteAccount(
-                              userData.uid,
-                              userData.financialInfo.accounts,
-                              "OtherAsset",
-                              accountKey
-                            );
-                          }}
-                          size="xs"
-                          variant="link"
-                        >
-                          Delete
-                        </Button>
-                      </CardBody>
-                    </Card>
+                      <a>
+                        <Card bgColor={"white"} justify="space-between">
+                          <CardBody>
+                            <Badge>OTHER ASSET</Badge>
+                            <Heading size="sm"> {currAccount.name} </Heading>
+                            <Stat>
+                              <StatLabel>Asset Value</StatLabel>
+                              <StatNumber>${currAccount.value}</StatNumber>
+                            </Stat>
+                            <Button
+                              onClick={() => {
+                                deleteAccount(
+                                  userData.uid,
+                                  userData.financialInfo.accounts,
+                                  "OtherAsset",
+                                  accountKey
+                                );
+                              }}
+                              size="xs"
+                              variant="link"
+                            >
+                              Delete
+                            </Button>
+                          </CardBody>
+                        </Card>
+                      </a>
+                    </Link>
                   );
                 }
               )}
@@ -278,53 +301,63 @@ export default function AccountsPage() {
                   const currAccount =
                     userData.financialInfo.accounts.creditCards[accountKey];
                   return (
-                    <Card
-                      bgColor={"white"}
+                    <Link
+                      href={`/app/accounts/${currAccount.name
+                        .toLowerCase()
+                        .split(" ")
+                        .join("-")}`}
                       key={accountKey}
-                      justify="space-between"
                     >
-                      <CardBody>
-                        <Badge>CREDIT CARD</Badge>
-                        <Heading size="sm"> {currAccount.name} </Heading>
-                        <Stat>
-                          <StatLabel>Amount Owed</StatLabel>
-                          <StatNumber>${currAccount.amountOwned}</StatNumber>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Account Interest Rate</StatLabel>
-                          <StatNumber>{currAccount.interestRate}%</StatNumber>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Next payment due</StatLabel>
-                          <StatLabel>
-                            {(currAccount.nextPaymentDate as Timestamp)
-                              .toDate()
-                              .toISOString()
-                              .substring(0, 10)}
-                          </StatLabel>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Next payment amount</StatLabel>
-                          <StatNumber>
-                            ${currAccount.nextPaymentAmount}
-                          </StatNumber>
-                        </Stat>
-                        <Button
-                          onClick={() => {
-                            deleteAccount(
-                              userData.uid,
-                              userData.financialInfo.accounts,
-                              "CreditCard",
-                              accountKey
-                            );
-                          }}
-                          size="xs"
-                          variant="link"
-                        >
-                          Delete
-                        </Button>
-                      </CardBody>
-                    </Card>
+                      <a>
+                        <Card bgColor={"white"} justify="space-between">
+                          <CardBody>
+                            <Badge>CREDIT CARD</Badge>
+                            <Heading size="sm"> {currAccount.name} </Heading>
+                            <Stat>
+                              <StatLabel>Amount Owed</StatLabel>
+                              <StatNumber>
+                                ${currAccount.amountOwned}
+                              </StatNumber>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Account Interest Rate</StatLabel>
+                              <StatNumber>
+                                {currAccount.interestRate}%
+                              </StatNumber>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Next payment due</StatLabel>
+                              <StatLabel>
+                                {(currAccount.nextPaymentDate as Timestamp)
+                                  .toDate()
+                                  .toISOString()
+                                  .substring(0, 10)}
+                              </StatLabel>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Next payment amount</StatLabel>
+                              <StatNumber>
+                                ${currAccount.nextPaymentAmount}
+                              </StatNumber>
+                            </Stat>
+                            <Button
+                              onClick={() => {
+                                deleteAccount(
+                                  userData.uid,
+                                  userData.financialInfo.accounts,
+                                  "CreditCard",
+                                  accountKey
+                                );
+                              }}
+                              size="xs"
+                              variant="link"
+                            >
+                              Delete
+                            </Button>
+                          </CardBody>
+                        </Card>
+                      </a>
+                    </Link>
                   );
                 }
               )}
@@ -335,53 +368,63 @@ export default function AccountsPage() {
                   const currAccount =
                     userData.financialInfo.accounts.loans[accountKey];
                   return (
-                    <Card
-                      bgColor={"white"}
+                    <Link
+                      href={`/app/accounts/${currAccount.name
+                        .toLowerCase()
+                        .split(" ")
+                        .join("-")}`}
                       key={accountKey}
-                      justify="space-between"
                     >
-                      <CardBody>
-                        <Badge>LOAN</Badge>
-                        <Heading size="sm">{currAccount.name}</Heading>
-                        <Stat>
-                          <StatLabel>Account Value</StatLabel>
-                          <StatNumber>
-                            ${currAccount.remainingAmount}
-                          </StatNumber>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Interest Rate</StatLabel>
-                          <StatNumber>{currAccount.interestRate}%</StatNumber>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Next payment due</StatLabel>
-                          <StatLabel>
-                            {(currAccount.paymentDate as Timestamp)
-                              .toDate()
-                              .toISOString()
-                              .substring(0, 10)}
-                          </StatLabel>
-                        </Stat>
-                        <Stat>
-                          <StatLabel>Minimum Payment</StatLabel>
-                          <StatNumber>${currAccount.minimumPayment}</StatNumber>
-                        </Stat>
-                        <Button
-                          onClick={() => {
-                            deleteAccount(
-                              userData.uid,
-                              userData.financialInfo.accounts,
-                              "Loan",
-                              accountKey
-                            );
-                          }}
-                          size="xs"
-                          variant="link"
-                        >
-                          Delete
-                        </Button>
-                      </CardBody>
-                    </Card>
+                      <a>
+                        <Card bgColor={"white"} justify="space-between">
+                          <CardBody>
+                            <Badge>LOAN</Badge>
+                            <Heading size="sm">{currAccount.name}</Heading>
+                            <Stat>
+                              <StatLabel>Account Value</StatLabel>
+                              <StatNumber>
+                                ${currAccount.remainingAmount}
+                              </StatNumber>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Interest Rate</StatLabel>
+                              <StatNumber>
+                                {currAccount.interestRate}%
+                              </StatNumber>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Next payment due</StatLabel>
+                              <StatLabel>
+                                {(currAccount.paymentDate as Timestamp)
+                                  .toDate()
+                                  .toISOString()
+                                  .substring(0, 10)}
+                              </StatLabel>
+                            </Stat>
+                            <Stat>
+                              <StatLabel>Minimum Payment</StatLabel>
+                              <StatNumber>
+                                ${currAccount.minimumPayment}
+                              </StatNumber>
+                            </Stat>
+                            <Button
+                              onClick={() => {
+                                deleteAccount(
+                                  userData.uid,
+                                  userData.financialInfo.accounts,
+                                  "Loan",
+                                  accountKey
+                                );
+                              }}
+                              size="xs"
+                              variant="link"
+                            >
+                              Delete
+                            </Button>
+                          </CardBody>
+                        </Card>
+                      </a>
+                    </Link>
                   );
                 }
               )}
